feat: type을 int에서 long으로 변경
- MemoryDb에서 JDBC Db로 변경할 예정이므로 long 타입으로 변경하였다.

+++ src/main/java/codesquad/springcafe/repository/ArticleRepository.java
@@ -4,7 +4,7 @@ import codesquad.springcafe.dto.Article;
 import java.util.List;
 
 public interface ArticleRepository {
-    int addArticle(Article article);
+    long addArticle(Article article);
 
     Article findArticleById(int id) throws Exception;
 
@@ -14,5 +14,5 @@ public interface ArticleRepository {
 
     List<Article> findAllArticle();
 
-    int increaseViewCount(Article article);
+    long increaseViewCount(Article article);
 }
